Some experimental code changes to try implementing unification
