+++ test/test_unification.cpp
@@ -2,36 +2,29 @@
 
 #include <boost/test/included/unit_test.hpp>
 
-//#include "unification.h"
+#include "Constant.h"
+#include "Variable.h"
+#include "unification.h"
+#include <iostream>
 
 using namespace std;
 
 BOOST_AUTO_TEST_CASE(test_unification) {
+    Constant c{"c"};
+    // Test whether an empty substitution returns failure
+    std::optional<Substitution> theta_failure = std::nullopt; 
+    BOOST_TEST(!(unify(c, c, theta_failure)).has_value());
 
-    BOOST_TEST(true);
-    // sample predicates for testing unification
-    //typedef boost::variant<client::ast::Entity, client::ast::Variable, Predicate> argument; // only boost::variant allows for recursion
-    //typedef vector<argument> v_args;
+    // Testing whether a non-empty substitution works
+    std::optional<Substitution> theta = Substitution();
+    theta.value()["c"] = Constant{"c_subst"};
+    auto subst = unify(c, c, theta);
 
-    //client::ast::Entity John("John");
-    //client::ast::Entity Richard("Richard");
-    //client::ast::Variable X("X");
+    BOOST_TEST(subst.has_value());
+    BOOST_TEST(subst.value()["c"].name == "c_subst");
 
-    //argument J1=John;
-    //argument R1=Richard;
-    //argument X1=X;
-
-    //v_args v1 (J1, R1);
-    //v_args v2 (J1, X1);
-
-    //KnowledgeBase::Predicate P1("Knows", v1);
-    //KnowledgeBase::Predicate P2("Knows", v2);
-
-    //// setting up to run unification 
-    //typedef unordered_map<string, string> subs;
-    //subs s1;
-
-    //s1 = KnowledgeBase.unification(P1, P2, s1);
-
-    //cout << s1;
+    // Check unification with variables
+    Variable v{"v"};
+    auto subst_2 = unify(v, c);
+    BOOST_TEST(subst_2.value()["v"].name == "c");
 }
